% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_community.R
\name{sample_community}
\alias{sample_community}
\title{Simulate sampling of a microbial community.}
\usage{
sample_community(community, nDNA = 20000, nRNA = 20000)
}
\arguments{
\item{community}{A tibble containing a microbial community. This will usually be the output of \code{generate_community}.}

\item{nDNA}{Number of rDNA gene sequences to simulate (i.e. rDNA read count)}

\item{nRNA}{Number of rRNA transcript sequences to simulate (i.e. rRNA read
count)}
}
\value{
A tibble containing the measured rDNA and rRNA abundances drawn from the
community. All OTUs in the input community, including those not detected,
will be included. The \code{phantom} column indicates 'phantom OTUs' that yielded
rRNA transcripts but not rDNA gene sequences.
}
\description{
\code{sample_community} simulates a sequencing-based sampling experiment on a
microbial community, with independent sequencing efforts for rDNA genes and
rRNA transcripts.  community. This allows for the effects of partial
sequencing to be evaluated.
}
\examples{

community <- generate_community(1000)
sample_community(community, nDNA = 500, nRNA = 500)

}
\seealso{
generate_community
}
